refactor(dto): convert money fields in request/response DTOs to BigDecimal
Prices and amounts on the wire are now BigDecimal across request and
response DTOs. Percentage fields (changePercent, unrealizedGainPercent,
etc.) stay Double — they are display ratios, not money.

+++ src/main/java/github/kaloyanov5/merkantil/dto/request/OrderRequest.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OrderRequest {
     @NotBlank(message = "Stock symbol is required")
@@ -20,5 +22,5 @@ public class OrderRequest {
     @NotNull(message = "Order type is required (MARKET/LIMIT)")
     private String orderType; // "MARKET" or "LIMIT"
 
-    private Double limitPrice; // Required if orderType is LIMIT
+    private BigDecimal limitPrice; // Required if orderType is LIMIT
 }
