+++ src/main/java/github/kaloyanov5/merkantil/dto/response/PortfolioGrowthResponse.java
@@ -2,15 +2,15 @@ package github.kaloyanov5.merkantil.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 /**
- * Portfolio value for a specific trading day
- * This represents reconstructed historical portfolio value
+ * Portfolio value for a specific trading day.
+ * This represents reconstructed historical portfolio value.
  */
 public record PortfolioGrowthResponse(
         @JsonFormat(pattern = "yyyy-MM-dd")
         LocalDate date,
-        Double value
+        BigDecimal value
 ) {}
-
