+++ src/main/java/github/kaloyanov5/merkantil/dto/response/PortfolioResponse.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,10 +14,10 @@ public class PortfolioResponse {
     private String symbol;
     private String stockName;
     private Integer quantity;
-    private Double averageBuyPrice;
-    private Double currentPrice;
-    private Double currentValue; // quantity × currentPrice
-    private Double totalCost; // quantity × averageBuyPrice
-    private Double unrealizedGain; // currentValue - totalCost
+    private BigDecimal averageBuyPrice;
+    private BigDecimal currentPrice;
+    private BigDecimal currentValue; // quantity × currentPrice
+    private BigDecimal totalCost; // quantity × averageBuyPrice
+    private BigDecimal unrealizedGain; // currentValue - totalCost
     private Double unrealizedGainPercent; // (unrealizedGain / totalCost) × 100
 }
